Scoreクラスありバージョン
